Deprecated {auto,component}_hybrid
- new cmake option
- compile targets for subdirectories
- more component extern templates
- deprecated subdir

Updated kokkos to include kokkosp_{begin,end}_fence

Disable IPO in testing

- IPO appears to be the reason that papi_array is failing with GCC 6

Formatting + forward decl

Update gotcha_tests.cpp

Update source/timemory/deprecated/CMakeLists.txt

Miscellaneous fixes

- Fixed shadows
- More modifications to avoid deprecated warnings
- account for T and T* in component_bundle

Update README.md

Update .travis.yml
