+++ cmake/Modules/Packages.cmake
@@ -461,6 +461,17 @@ endif()
 
 #----------------------------------------------------------------------------------------#
 #
+#                           Deprecated code
+#
+#----------------------------------------------------------------------------------------#
+
+
+if(TIMEMORY_USE_DEPRECATED)
+    timemory_target_compile_definitions(timemory-headers INTERFACE TIMEMORY_USE_DEPRECATED)
+endif()
+
+#----------------------------------------------------------------------------------------#
+#
 #                           Cereal (serialization library)
 #
 #----------------------------------------------------------------------------------------#
