+++ source/timemory/components/extern/common.hpp
@@ -33,13 +33,17 @@
 #include "timemory/components/macros.hpp"
 #include "timemory/components/opaque/declaration.hpp"
 #include "timemory/environment/declaration.hpp"
+#include "timemory/operations/types/add_secondary.hpp"
+#include "timemory/operations/types/add_statistics.hpp"
 #include "timemory/operations/types/assemble.hpp"
+#include "timemory/operations/types/audit.hpp"
 #include "timemory/operations/types/cache.hpp"
 #include "timemory/operations/types/cleanup.hpp"
 #include "timemory/operations/types/construct.hpp"
 #include "timemory/operations/types/copy.hpp"
 #include "timemory/operations/types/derive.hpp"
 #include "timemory/operations/types/echo_measurement.hpp"
+#include "timemory/operations/types/finalize/ctest_notes.hpp"
 #include "timemory/operations/types/finalize/dmp_get.hpp"
 #include "timemory/operations/types/finalize/flamegraph.hpp"
 #include "timemory/operations/types/finalize/get.hpp"
@@ -47,9 +51,12 @@
 #include "timemory/operations/types/finalize/mpi_get.hpp"
 #include "timemory/operations/types/finalize/print.hpp"
 #include "timemory/operations/types/finalize/upc_get.hpp"
+#include "timemory/operations/types/fini.hpp"
 #include "timemory/operations/types/fini_storage.hpp"
 #include "timemory/operations/types/get.hpp"
+#include "timemory/operations/types/init.hpp"
 #include "timemory/operations/types/init_storage.hpp"
+#include "timemory/operations/types/math.hpp"
 #include "timemory/operations/types/node.hpp"
 #include "timemory/operations/types/print.hpp"
 #include "timemory/operations/types/print_header.hpp"
@@ -60,6 +67,7 @@
 #include "timemory/operations/types/serialization.hpp"
 #include "timemory/operations/types/set.hpp"
 #include "timemory/operations/types/storage_initializer.hpp"
+#include "timemory/operations/types/store.hpp"
 #include "timemory/plotting/declaration.hpp"
 #include "timemory/settings/declaration.hpp"
 #include "timemory/storage/declaration.hpp"
