+++ source/timemory/deprecated/CMakeLists.txt
@@ -1,0 +1,18 @@
+
+file(GLOB_RECURSE header_files ${CMAKE_CURRENT_SOURCE_DIR}/*.hpp)
+file(GLOB_RECURSE source_files ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
+
+if(TIMEMORY_USE_DEPRECATED)
+    build_intermediate_library(
+        USE_INTERFACE
+        INSTALL_SOURCE
+        NAME                deprecated
+        TARGET              deprecated
+        CATEGORY            GLOBAL
+        FOLDER              components
+        HEADERS             ${header_files}
+        SOURCES             ${source_files}
+        PROPERTY_DEPENDS    GLOBAL)
+else()
+    timemory_install_header_files(${header_files} ${source_files})
+endif()
