+++ source/timemory/ert/counter.hpp
@@ -68,7 +68,6 @@ template <typename DeviceT, typename Tp, typename Counter>
 class counter
 {
 public:
-    using string_t      = std::string;
     using mutex_t       = std::recursive_mutex;
     using lock_t        = std::unique_lock<mutex_t>;
     using counter_type  = Counter;
