+++ source/timemory/storage/definition.hpp
@@ -423,7 +423,8 @@ storage<Type, true>::finalize()
     manager::instance()->is_finalizing(true);
 
     using fini_t = operation::fini<Type>;
-    auto upcast  = static_cast<tim::storage<Type>*>(this);
+    using ValueT = typename trait::collects_data<Type>::type;
+    auto upcast  = static_cast<tim::storage<Type, ValueT>*>(this);
 
     if(m_thread_init)
         fini_t(upcast, operation::mode_constant<operation::fini_mode::thread>{});
@@ -466,7 +467,8 @@ storage<Type, true>::global_init()
             if(m_is_master)
             {
                 using init_t = operation::init<Type>;
-                auto upcast  = static_cast<tim::storage<Type>*>(this);
+                using ValueT = typename trait::collects_data<Type>::type;
+                auto upcast  = static_cast<tim::storage<Type, ValueT>*>(this);
                 init_t(upcast, operation::mode_constant<operation::init_mode::global>{});
             }
             return m_global_init;
@@ -488,7 +490,8 @@ storage<Type, true>::thread_init()
             bool _global_init = global_init();
             consume_parameters(_global_init);
             using init_t = operation::init<Type>;
-            auto upcast  = static_cast<tim::storage<Type>*>(this);
+            using ValueT = typename trait::collects_data<Type>::type;
+            auto upcast  = static_cast<tim::storage<Type, ValueT>*>(this);
             init_t(upcast, operation::mode_constant<operation::init_mode::thread>{});
             return m_thread_init;
         }();
@@ -1089,8 +1092,9 @@ storage<Type, false>::initialize()
 
     m_initialized = true;
 
-    auto upcast  = static_cast<tim::storage<Type>*>(this);
     using init_t = operation::init<Type>;
+    using ValueT = typename trait::collects_data<Type>::type;
+    auto upcast  = static_cast<tim::storage<Type, ValueT>*>(this);
 
     if(!m_is_master)
     {
@@ -1119,7 +1123,8 @@ storage<Type, false>::finalize()
         printf("[%s]> finalizing...\n", m_label.c_str());
 
     using fini_t = operation::fini<Type>;
-    auto upcast  = static_cast<tim::storage<Type>*>(this);
+    using ValueT = typename trait::collects_data<Type>::type;
+    auto upcast  = static_cast<tim::storage<Type, ValueT>*>(this);
 
     m_finalized = true;
     manager::instance()->is_finalizing(true);
