+++ source/timemory/tpls.hpp
@@ -22,29 +22,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-/**
- * \file timemory/containers/definition.hpp
- * \brief The definitions for the types in containers
- */
-
 #pragma once
 
-#include "timemory/containers/auto_timer.hpp"
-#include "timemory/containers/auto_user_bundle.hpp"
-#include "timemory/containers/declaration.hpp"
-#include "timemory/containers/macros.hpp"
-#include "timemory/containers/types.hpp"
-#include "timemory/operations/definition.hpp"
-#include "timemory/storage/definition.hpp"
-#include "timemory/variadic/auto_bundle.hpp"
-#include "timemory/variadic/auto_list.hpp"
-#include "timemory/variadic/auto_tuple.hpp"
-#include "timemory/variadic/component_bundle.hpp"
-#include "timemory/variadic/component_list.hpp"
-#include "timemory/variadic/component_tuple.hpp"
-#include "timemory/variadic/lightweight_tuple.hpp"
-
-#if defined(TIMEMORY_USE_DEPRECATED)
-#    include "timemory/variadic/auto_hybrid.hpp"
-#    include "timemory/variadic/component_hybrid.hpp"
-#endif
+#include "timemory/tpls/cereal/cereal.hpp"
