+++ source/timemory/variadic/definition.hpp
@@ -28,21 +28,27 @@
 #include "timemory/variadic/types.hpp"
 //
 #include "timemory/variadic/auto_bundle.hpp"
-#include "timemory/variadic/auto_hybrid.hpp"
 #include "timemory/variadic/auto_list.hpp"
 #include "timemory/variadic/auto_tuple.hpp"
 #include "timemory/variadic/base_bundle.hpp"
 #include "timemory/variadic/component_bundle.hpp"
-#include "timemory/variadic/component_hybrid.hpp"
 #include "timemory/variadic/component_list.hpp"
 #include "timemory/variadic/component_tuple.hpp"
 #include "timemory/variadic/functional.hpp"
 #include "timemory/variadic/lightweight_tuple.hpp"
 //
+#if defined(TIMEMORY_USE_DEPRECATED)
+#    include "timemory/variadic/auto_hybrid.hpp"
+#    include "timemory/variadic/component_hybrid.hpp"
+#endif
+//
 //  Implementations
 //
 #include "timemory/variadic/component_bundle.cpp"
-#include "timemory/variadic/component_hybrid.cpp"
 #include "timemory/variadic/component_list.cpp"
 #include "timemory/variadic/component_tuple.cpp"
 #include "timemory/variadic/lightweight_tuple.cpp"
+//
+#if defined(TIMEMORY_USE_DEPRECATED)
+#    include "timemory/deprecated/component_hybrid.cpp"
+#endif
